refactor: post util 함수로 분리
