+++ src/apis/orderProducts.ts
@@ -1,7 +1,7 @@
 import { servers } from '../constants/server';
+import { fetchData, postData } from '../utils/apiUtils';
 import type { OrderedData, OrderedProduct } from '../types/product';
 import type { HostNameType } from '../types/server';
-import { fetchData } from '../utils/apiUtils';
 
 const email = process.env.REACT_APP_EMAIL;
 const password = process.env.REACT_APP_PASSWORD;
@@ -29,31 +29,16 @@ export const orderApi = async (hostName: HostNameType) => {
         Authorization: `Basic ${base64}`,
       },
     });
-
     return response;
   };
 
   const postOrderProduct = async (orderData: OrderedData) => {
-    const response = await fetch(URL, {
-      method: 'POST',
-      headers: {
-        Authorization: `Basic ${base64}`,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(orderData),
-    });
+    const headers = {
+      Authorization: `Basic ${base64}`,
+    };
 
-    if (!response.ok) {
-      throw new Error(response.status.toString());
-    }
-
-    const location = response.headers.get('location');
-
-    if (location !== null) {
-      const lastSlashIndex = location.lastIndexOf('/');
-      const orderId = location.slice(lastSlashIndex + 1);
-      return orderId;
-    }
+    const response = await postData(URL, headers, { orderData });
+    return response;
   };
 
   return {
