+++ src/types/fetch.ts
@@ -1,0 +1,6 @@
+import type { OrderedData } from './product';
+
+export interface postDataType {
+  productId?: number;
+  orderData?: OrderedData;
+}
